V44 - Geometry data is now parsed for woonplaatsen, panden, standplaatsen and ligplaatsen. - config.py: New option "parse_geometries" - Improved tests - Small modifications

+++ utils_sqlite_shrink.py
@@ -1,10 +1,12 @@
+import config
 import utils
 from database_sqlite import DatabaseSqlite
 
 db_sqlite = DatabaseSqlite()
 
-utils.print_log('Delete full BAG tables')
-db_sqlite.delete_no_longer_needed_bag_tables()
+if config.create_adressen_table and config.delete_no_longer_needed_bag_tables:
+    utils.print_log('Delete full BAG tables')
+    db_sqlite.delete_no_longer_needed_bag_tables()
 
 utils.print_log('cleaning up: vacuum')
 db_sqlite.commit()
